Crate subxt_metadata

A representation of the metadata provided by a substrate based node. This representation is optimized to be used by Subxt and related crates, and is independent of the different versions of metadata that can be provided from a node.

Typically, this will be constructed by either:

Calling Metadata::decode() given some metadata bytes obtained from a node (this uses codec::Decode).

Obtaining frame_metadata::RuntimeMetadataPrefixed, and then using .try_into() to convert it into Metadata.

Structs

ConstantMetadata

Metadata for a single constant.

CustomMetadata

Metadata of custom types with custom values, basically the same as frame_metadata::v15::CustomMetadata<PortableForm>>.

CustomValueMetadata

Basically the same as frame_metadata::v15::CustomValueMetadata<PortableForm>>, but borrowed.

ExtrinsicMetadata

Metadata for the extrinsic type.

Metadata

Node metadata. This can be constructed by providing some compatible frame_metadata which is then decoded into this. We aim to preserve all of the existing information in the incoming metadata while optimizing the format a little for Subxt’s use cases.

MetadataHasher

Obtain a hash representation of our metadata or some part of it. This is obtained by calling crate::Metadata::hasher().

OuterEnumsMetadata

Metadata for the outer enums.

PalletMetadata

Metadata for a specific pallet.

RuntimeApiMetadata

Metadata for the available runtime APIs.

RuntimeApiMethodMetadata

Metadata for a single runtime API method.

RuntimeApiMethodParamMetadata

Metadata for a single input parameter to a runtime API method.

SignedExtensionMetadata

Metadata for the signed extensions used by extrinsics.

StorageEntryMetadata

Metadata for a single storage entry.

StorageMetadata

Metadata for the storage entries in a pallet.

Enums

StorageEntryModifier

Is the storage entry optional, or does it have a default value.

StorageEntryType

The type of a storage entry.

StorageHasher

Hasher used by storage maps.

TryFromError

An error emitted if something goes wrong converting frame_metadata types into crate::Metadata.